Define common QueryBuilder interface.
The interface should be "implemented" by all classes that are query builders or act as one by proxying calls. I am writing "implemented" in quotes because the interface does not define any methods. Thus it is not strictly an interface, but rather a type.

Why the interface then, if it doesn't define any methods? Because it would be awesome to have a way of type-hinting arguments and return types of methods that require or return query builders, but accept all three types (Database\Builder, Eloquent\Builder and Relations\Relation). Right now, the only option is to not use any type hints at all, if a method works on more that one of those three types.